[rename] 각 테스트에 대한 Template를 만들어서 코드 이동
