+++ src/test/java/springJpaBoard/Board/BoardApplicationTests.java
@@ -1,13 +1,8 @@
 package springJpaBoard.Board;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class BoardApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
 
 }
